Add timers and signals, support testing of kernel
* added: usleep() now works. Threads can call usleep() and be suspended
  and then resumed once timer expires.
* added: setitimer() half-works. Threads can call setitimer(). Periodic
  timer will be scheduled, but delivery of SIGALRM is not verified as
  working yet.
* added: signal() and kill() implemented, untested. Thread can opt for
  signal handler routine registration. Thread can also emit signal for
  other thread. Signals STOP, CONT, KILL, SEGV are uncatchable and have
  immediate effect of stopping, resuming, killing and killing a thread.
  Other signals are catchable and cause signal handler to be injected
  into thread execution. If thread was stopped before, it will also be
  resumed up.
* added: elementary support for kernel testing. Some types which won't
  play nicely with hosted code on AMD64 / AArch64 were changed to be of
  native width. Note: this most probably breaks `make test` as of now.
  Take the red pill and fix testing ASAP.

+++ include/cmrx/intrinsics.h
@@ -66,7 +66,7 @@ __attribute__((always_inline)) static inline void * __get_LR(void)
 /** Set value of process SP
  * @param stack_top new top of application stack
  */
-__attribute__((always_inline)) static inline void __set_PSP(uint32_t * stack_top)
+__attribute__((always_inline)) static inline void __set_PSP(unsigned long * stack_top)
 {
 	asm volatile(
 			".syntax unified\n\t"
