+++ include/cmrx/os/sched/stack.h
@@ -13,7 +13,7 @@
  */
 struct OS_stack_t {
 	/** Thread stacks. */
-	uint32_t stacks[OS_STACKS][OS_STACK_DWORD];
+	unsigned long stacks[OS_STACKS][OS_STACK_DWORD];
 
 	/** Information about stack allocation. If n-th bit is set, 
 	 * then n-th stack is allocated. Otherwise it is available. */
